Changed grad kernel type **AGAIN**. I will have to squash all that stuff

+++ train.py
@@ -476,7 +476,7 @@ def DSSIM_MSE():
 
 def mse_plus_grad(alpha=0.6):
     def loss(y_true, y_pred):
-        grad_kernel = K.variable([[0, -1, 0],[-1, 0, 1],[0, 1, 0]], dtype='int64')
+        grad_kernel = K.variable([[0, -1, 0],[-1, 0, 1],[0, 1, 0]], dtype=y_true.dtype)
         grad_true = K.conv2d(y_true, grad_kernel, padding='same', strides=(1,1))
         grad_pred = K.conv2d(y_pred, grad_kernel, padding='same', strides=(1,1))
         return alpha*keras.losses.mean_squared_error(y_true,y_pred) + (1-alpha)*keras.losses.mean_squared_error(grad_true,grad_pred)
